feat: 게시글 생성 DTO 추가 및 변경 (#4)
